api: patch service -> LoadBalancer
For API operator charms, patch the Juju configured service to
a) configure a port and b) switch the type of the service to
LoadBalancer.

This ensures that external access to the API service is configured
in the charm, rather than being a post deployment step.

As the ingress address of the loadbalancer can be discovered via
the K8S API, use this for the Public URL for the service.