Remove more use of stored state
service_ready

Refactor Pebble container handlers to directly query the container
when assessing whether the container service is ready or not.

Updated test harness code to use super calls for {start,stop}_service
to ensure that service state modelling is correct; this also required
a tweak to start all to ensure that services that are already started
do not get started again (testing harness throws a state error).

config_pushed

Not actually used anywhere so dropped.

Change-Id: I5fc7bb5c5e1a41341856d6ee226c9bc5c008f0f9