Add run_once_per_unit decorator
This change adds a decorator `run_once_per_unit` which can be used
to limit running a method to once per instantiation of a unit.

This decorator relies on using ops framework BoundStoredState with
use_juju_for_storage=False as the storage must be cleared when
the unit is recycled. (use_juju_for_storage=True is deprecated
by the ops framework so should not be used anyway). To enforce
this the charm will throw an exeption if use_juju_for_storage=True
all existing sunbeam charms will need updating to accomadate
this.

There is a follow up patch which adds `run_once_per_app` which is
a decorator that runs once per version of the application (so only
the leader runs it) but this is more complicated and less useful
(imo) so I will propose it seperatly.

Change-Id: I5109f1a462208eba5bfa7dd484448511d686bf19