Refactor bootstrap
Break up the configure_charm method into distinct phases each
with a dedicated method. This makes it much easier for a charm
specialising this class to customise and individual phase
without having to duplicate code.

To enable this the configure_charm method has been adapted to use
the guard function. This allows any method within the guard
context manager to throw a guard exception to halt the
configuration process and have the issue reflected in the
units workload status. This is more flexable approach then
the old method of having configuration steps return a
boolean.

Change-Id: I6b46bae1dc700f5e8b597374c660c0af9c069172