Use public method for url retrieval for traefik
Due to a bug in the public url method sunbeam was using the
private _get_url_from_relation_data. This ignores stored state
and interogates the relation data directly. This causes a
race condition when new traefik units are coming up. The bug in
url property is fixed in v2 of the traefik interface so switch
to using that instead.

Closes-Bug: #2036990
Change-Id: I883ca62f8b6e072648bf69d9481bd6c11014ae6b